changed the landing_page file name to camel stile | removed the comments | renamed functions
